refactor: Split component A into a separate file for better organization
