+++ my-git/src/App.js
@@ -1,29 +1,10 @@
 import React, { useState } from "react";
+import A from "./A";
 
 function App() {
   return (
     <div>
       <A />
-    </div>
-  );
-}
-
-function A() {
-  const [showB, setShowB] = useState(false);
-  const [showC, setShowC] = useState(false);
-  const [text, setText] = useState("");
-
-  const handleInputChange = (event) => {
-    setText(event.target.value);
-    setShowB(true);
-    setShowC(true);
-  };
-
-  return (
-    <div className="App">
-      <input type="text" onChange={handleInputChange} />
-      {showB && <B text={text} />}
-      {showC && <C text={text} />}
     </div>
   );
 }
